Add debug option to play video and videorate element
videorate makes sure the video is saved as it would be played out, i.e.
adding duplicate frames where frames are dropped instead of just
skipping over missing frames. This ensures the video will have roughly
the same length and number of frames as the original. It is achieved by
adding as many duplicate frames as required to keep up the framerate.